requirements.txt
crewai==0.70.1
google-generativeai==0.8.3
scikit-learn==1.5.1
spacy==3.7.4
sumy==0.11.0
pdfminer.six==20231228
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==1.1.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
Pillow==10.4.0
nltk==3.9.1